Define oob_config_dir env var and interpolation
With this patch, path references in the config file
can refer to %(oob_config_dir)s which resolves to the
directory where the config file resides.

Additionally, scripts will find $oob_config_dir defined
in the environment.

For example this will find a custom script in the same dir
as the config file:

  [custom_scripts]
  custom_script_1=%(oob_config_dir)s/myscript.sh

and in a custom script, this will apply a patch in the same
directory as the config file:

  patch $INSTALL_DIR/etc/foo < $oob_config_dir/foo.patch