Define oob_config_dir env var and interpolation
Path references in the config file can now refer to %(oob_config_dir)s
which resolves to the directory where the config file resides.

Additionally, scripts will find $oob_config_dir defined in the environment.

For example this will find a custom script in the same dir as the config
file:

  [custom_scripts]
  custom_script_1=%(oob_config_dir)s/myscript.sh

In a custom script, this will apply a patch in the same directory as the
config file:

	patch $INSTALL_DIR/etc/foo < $oob_config_dir/foo.patch

[dsd: minor tweak, add documentation]