fix serial console (#10845, #11008, #11044, #10669, #11045)
- remove agetty from /etc/rc.local,
- move existing unused agetty from ttyS0 to ttyS2, (adding -L to agetty),
- re-enable plymouth,

Tested on os19 with XO-1.75 A2 and A3.

The serial port we use does not have modem control signals and
apparently should not be asked to do carrier detect.  Exactly why is
not clear, but probably not important.

The temporary agetty we had in /etc/rc.local was doing carrier
detect, resulting in several hangs:

- #11008 XO-1.75 os11 reboot hangs, caused by plymouth trying to open
  the serial console,

- #10669 XO-1.75 os3 shutdown doesn't (plymouth stall), kernel panic,

- #10845 XO-1.75 os6 serial shell chvt hangs on open /proc/self/fd/0,
  which was effectively another open of the serial console,

This patch also removes a superfluous agetty process (#11045), and
correctly respawns the serial console process (#11044).

The problem was affecting olpc-runin-tests use of agetty debug
consoles.